change convolution function to take floats and fix image border

+++ include/ImageProcessing.h
@@ -2,8 +2,7 @@
 #include <cstdint>
 #include <vector>
 
-float* convolveImageKernel(float *image, int width, int height, std::vector<std::vector<int>> kernel);
-
+float* convolveImageKernel(float *image, int width, int height, std::vector<std::vector<float>> kernel);
 float* harrisCornerDetector(float *image, int width, int height, int blockSize, float sensitivity);
 void shiTomasiCornerDetector();
